filter 필드의 데이터 크기(byte size)가 정수형을 넘어설때 에러나는 버그수정.

+++ core/src/main/java/org/fastcatsearch/ir/search/ReferenceableIndexReader.java
@@ -69,7 +69,7 @@ public abstract class ReferenceableIndexReader implements ReferenceableReader {
 				dataRef.init(0);
 			}
 		}else{
-			int pos = dataSize * docNo;
+            long pos = ((long) dataSize) * docNo;
 			dataInput.seek(pos);
 			//이미 input의 position을 움직여 놓았으므로 더이상 아무것도 하지 않는다.
 			dataInput.readBytes(dataRef.bytesRef().bytes, 0, dataSize);
